+++ electron/src/frontend/components/ui/codeblock.tsx
@@ -4,7 +4,7 @@ import { FC, memo } from 'react'
 import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter'
 import { coldarkDark } from 'react-syntax-highlighter/dist/cjs/styles/prism'
 
-import { useCopyToClipboard } from '@/lib/hooks/chat.use-copy-to-clipboard'
+import { useCopyToClipboard } from '@/lib/hooks/use-copy-to-clipboard'
 import { Copy, CopyCheck, Download } from 'lucide-react'
 import { Button } from '@/components/ui/button'
