+++ electron/src/frontend/lib/hooks/use-copy-to-clipboard.ts
